chore(ci): Sync logic between travis/npm/grunt
The industry practice is generally for projects that use Node.js,
for most (if not, all) tests to be run via 'npm test'. A command such as
'npm run lint' is expected to be a subset of that.

That is exactly what was the case here, but in a not-so-obvious way.

* Remove superfluous 'yarn lint' script in addition to 'yarn test',
  which already did the same (yarn test > scripts.test >
  grunt default > eslint + grunt test).

* Move the running of eslint as part of "test" further down the
  stack (into 'grunt test' instead of joining only as part of
  grunt [default]).

* Remove unused pgkFile property.

* Rename gruntfile.js to Gruntfile.js, matching the current
  practice and expectation of users and contributors.